empty bloom no alloc #786
