+++ src/Nethermind/Nethermind.Network.Test/P2P/Subprotocols/Eth/NewBlockMessageSerializerTests.cs
@@ -16,7 +16,6 @@
  * along with the Nethermind. If not, see <http://www.gnu.org/licenses/>.
  */
 
-using DotNetty.Buffers;
 using Nethermind.Core;
 using Nethermind.Core.Extensions;
 using Nethermind.Core.Test.Builders;
@@ -58,47 +57,6 @@ namespace Nethermind.Network.Test.P2P.Subprotocols.Eth
             byte[] bytes = serializer.Serialize(newBlockMessage);
             byte[] expectedBytes = Bytes.FromHexString("f90243f9023ff901f9a0ff483e972a04a9a62bb4b7d04ae403c615604e4090521ecc5bb7af67f71be09ca01dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347940000000000000000000000000000000000000000a056e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421a056e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421a056e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421b9010000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000830f424080833d090080830f424083010203a02ba5557a4c62a513c7e56d1bf13373e0da6bec016755483e91589fe1c6d212e28800000000000003e8f840df80018252089400000000000000000000000000000000000000000180808080df80018252089400000000000000000000000000000000000000000180808080c080");
             
-            Assert.AreEqual(expectedBytes.ToHexString(), bytes.ToHexString(), "bytes");
-        }
-    }
-    
-    [TestFixture]
-    public class ZeroNewBlockMessageSerializerTests
-    {
-        [Test]
-        public void Roundtrip()
-        {
-            Transaction a = Build.A.Transaction.TestObject;
-            Transaction b = Build.A.Transaction.TestObject;
-            Block block = Build.A.Block.WithTransactions(a, b).TestObject;
-            NewBlockMessage newBlockMessage = new NewBlockMessage();
-            newBlockMessage.Block = block;
-            
-            ZeroNewBlockMessageSerializer serializer = new ZeroNewBlockMessageSerializer();
-            IByteBuffer byteBuffer = PooledByteBufferAllocator.Default.Buffer(1024);
-            serializer.Serialize(byteBuffer, newBlockMessage);
-            byte[] expectedBytes = Bytes.FromHexString("07f90243f9023ff901f9a0ff483e972a04a9a62bb4b7d04ae403c615604e4090521ecc5bb7af67f71be09ca01dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347940000000000000000000000000000000000000000a056e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421a056e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421a056e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421b9010000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000830f424080833d090080830f424083010203a02ba5557a4c62a513c7e56d1bf13373e0da6bec016755483e91589fe1c6d212e28800000000000003e8f840df80018252089400000000000000000000000000000000000000000180808080df80018252089400000000000000000000000000000000000000000180808080c080");
-            byte[] bytes = byteBuffer.ReadAllBytes();
-            
-            byteBuffer.ReadBytes(bytes);
-            Assert.AreEqual(expectedBytes.ToHexString(), bytes.ToHexString(), "bytes");
-        }
-        
-        [Test]
-        public void Roundtrip2()
-        {
-            Transaction a = Build.A.Transaction.TestObject;
-            Transaction b = Build.A.Transaction.TestObject;
-            Block block = Build.A.Block.WithTransactions(a, b).TestObject;
-            NewBlockMessage newBlockMessage = new NewBlockMessage();
-            newBlockMessage.Block = block;
-            
-            ZeroNewBlockMessageSerializer serializer = new ZeroNewBlockMessageSerializer();
-            IByteBuffer byteBuffer = PooledByteBufferAllocator.Default.Buffer(1024);
-            serializer.Serialize(byteBuffer, newBlockMessage);
-            byte[] expectedBytes = Bytes.FromHexString("07f90243f9023ff901f9a0ff483e972a04a9a62bb4b7d04ae403c615604e4090521ecc5bb7af67f71be09ca01dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347940000000000000000000000000000000000000000a056e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421a056e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421a056e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421b9010000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000830f424080833d090080830f424083010203a02ba5557a4c62a513c7e56d1bf13373e0da6bec016755483e91589fe1c6d212e28800000000000003e8f840df80018252089400000000000000000000000000000000000000000180808080df80018252089400000000000000000000000000000000000000000180808080c080");
-            byte[] bytes = byteBuffer.ReadAllBytes();
-
             Assert.AreEqual(expectedBytes.ToHexString(), bytes.ToHexString(), "bytes");
         }
     }
